backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.3
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
